requirements.txt
streamlit==1.30.0
openai==1.6.1
langchain==0.1.9
langchain-community==0.0.25  #13 Updated to match langchain's requirement
faiss-cpu==1.7.4
python-dotenv==1.0.0
#tensorflow
transformers==4.35.2  # Compatible with Keras <3
torch==2.1.0   # Ensures compatibility with Transformers
numpy==1.23.5  # Prevents FAISS compatibility issues
sympy==1.12.0